+++ src/drives/SingleTapeDrive.java
@@ -3,7 +3,8 @@ package drives;
 import java.util.ArrayList;
 import java.util.List;
 
-import program.movements.Movement;
+import machine.Movement;
+
 
 import exceptions.MachineStoppedException;
