+++ src/machine/Machine.java
@@ -8,7 +8,6 @@ import org.javatuples.Triplet;
 
 import program.Program;
 import program.loader.ProgramLoader;
-import program.movements.Movement;
 
 
 import drives.Drive;
